+++ src/foundations/fonts.ts
@@ -1,4 +1,4 @@
 export default {
-  heading: 'InterVariable, -apple-system, system-ui, sans-serif',
-  body: 'InterVariable, -apple-system, system-ui, sans-serif',
+  heading: '"Inter Variable", -apple-system, system-ui, sans-serif',
+  body: '"Inter Variable", -apple-system, system-ui, sans-serif',
 }